remove Delimi and Download dependencies
