+++ Project.toml
@@ -1,11 +1,9 @@
 name = "TransitionsInTimeseries"
 uuid = "5f5b98ec-1183-43e0-887a-12fdc55c52f7"
 authors = ["Jan Swierczek-Jereczek <jan.jereczek@gmail.com>", "George Datseris <datseris.george@gmail.com>"]
-version = "0.1.4"
+version = "0.2.0"
 
 [deps]
-DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
-Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 HypothesisTests = "09f84164-cd44-5f33-b23f-e6b0d136a0d5"
 InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
@@ -24,8 +22,6 @@ Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 TransitionVisualizations = "Makie"
 
 [compat]
-DelimitedFiles = "1"
-Downloads = "1"
 FFTW = "^1.6"
 HypothesisTests = "0.11"
 InteractiveUtils = "1"
